Established basis of the navigation bar
Established an additional file structure and created the new layout establishment of the Homepage. Homepage is now exported using the react-router-dom and the Layout is now established for additional structure to the webpage. Here is the heirarchy:
App.js --> Home --> Layout --> Navbar

Any other bars will likely be inputted into the Layout.tsx file and any other elementes will be displayed as children to the div in the layout.tsx file below the Navbar