Major changes to Navbar and Components
Connected all of the components to the constants and so now constants are being displayed. Navbar is now sticky and at the top, footer is at the bottom and center aligned. Components occupy an entire page instead of just their respective content boundaries. Additional types were added to the experience and project links to comply with typescript checking.
Files edited:
new file:   notes.txt