Revised Navigation Bar & Landing Page
Added a temporary gradient to the background of the landing page, changed the fonts on the text. Centered the elements on the landing page and added an image that I may or may not keep on that landing page. I think I am right now leaning towards moving that image to the about section instead. I also have tried a few implementations of adding transitions/animations to the fonts of my landing page but have not been too successful: namely I used react-reveal and react-awesome-reveal. I also had trouble with react-text-transition. This is still in the works and is something I will get to tommorow. I am hoping to finish this navbar and landing page tomorrow as well as the About Page tomorrow.

Current Sprint Plan for the week is as follows:
Wednesday - 6/28
- Styling of Navbar, Landing Page & About Page
Thursday - 6/29
- Styling of Experience, Projects Page
Friday - 6/30
- Mobile Screen sizing adjustments, deployment